pyproject.toml
[tool.poetry]
name = "core-backend"
version = "0.1.0"
description = ""
authors = ["Alfur <zbsrebyata@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
isort = "^5.13.2"
black = "^24.3.0"
flake8 = "^7.0.0"
drf-spectacular = "^0.27.1"
djangorestframework-simplejwt = "^5.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
